synch code from 2394

+++ cpp/src/category/numeric_category.inl
@@ -14,7 +14,6 @@
 * limitations under the License.
 */
 
-#include "../include/numeric_category.h"
 #include <thrust/device_vector.h>
 #include <thrust/host_vector.h>
 #include <thrust/for_each.h>
@@ -29,6 +28,8 @@
 #include <thrust/tabulate.h>
 #include <rmm/rmm.h>
 #include <rmm/thrust_rmm_allocator.h>
+
+#include "../include/numeric_category.h"
 
 
 #define BYTES_FROM_BITS(c) ((c+7)/8)
@@ -708,7 +709,7 @@ numeric_category<T>* numeric_category<T>::set_keys( const T* items, size_t count
                     return lhs_null == rhs_null;
                 return (d_both_keys[lhs] == d_both_keys[rhs]);
             });
-    int copy_count = d_copy_end - d_map_indexes;
+    size_t copy_count = d_copy_end - d_map_indexes;
     if( copy_count < both_count )
     {   // if keys are removed, we need new keyset; the gather()s here will select the remaining keys
         rmm::device_vector<int> copy_indexes(copy_count);
